CN-7176 Adding principal alert conditions
